refactor: エンジンマニフェストに `model` モジュールを追加
